Set Speed seperated, need to change controls to the POV
